refactor: separated vault subject into it's own sub router
